+++ src/hooks/HookedGameManager.cpp
@@ -39,7 +39,6 @@ CCTexture2D* HookedGameManager::loadIcon(int iconID, int iconType, int _idk) {
 void HookedGameManager::reportPercentageForLevel(int levelID, int percent, bool isPractice) {
 
     GameManager::reportPercentageForLevel(levelID, percent, isPractice);
-    log::info("hi: {}, {}, {}", levelID, percent, isPractice);
 
     if(levelID == 10 && !isPractice && percent >= 60 && percent < 100)
         GameManager::sharedState()->reportAchievementWithID("geometry.ach.lunar.percent01", 100, false);