+++ src/hooks/HookedGameStatsManager.cpp
@@ -16,9 +16,6 @@ void HookedGameStatsManager::checkCoinAchievement(GJGameLevel* level) {
     int coinPercent = (getCollectedCoinsForLevel(level) / 3.f) * 100.f;
     auto achievementID = fmt::format("geometry.ach.lunar.levelcoin{:02}", level->m_levelID.value());
 
-    log::info("coinPercent: {}, achID: {}", coinPercent, achievementID);
-
-
     if(level->m_levelType == GJLevelType::Local) {
         GM->reportAchievementWithID(achievementID.c_str(), coinPercent, false);
     }